rename redacted module to unredacted
